Introduce new structures
Introduce new structures for job templates, machines, product,
testsuite. Also move job and worker to their own definition file.
