+++ machine.go
@@ -1,0 +1,8 @@
+package gopenqa
+
+/* Machine type */
+type Machine struct {
+	ID      int    `json: "id"`
+	Backend string `json: "backend"`
+	Name    string `json: "name"`
+}
